Move `usUtils` to foundation.
