+++ yarn-project/circuits.js/src/index.ts
@@ -2,4 +2,3 @@ export * from './structs/index.js';
 export * from './wasm/index.js';
 export * from './kernel/index.js';
 export * from './rollup/index.js';
-export * from './utils/jsUtils.js';
