+++ yarn-project/circuits.js/src/structs/kernel/previous_kernel_data.ts
@@ -1,7 +1,6 @@
 import { Fr } from '@aztec/foundation/fields';
 import times from 'lodash.times';
 import { CircuitsWasm, getDummyPreviousKernelData } from '../../index.js';
-import { assertLength } from '../../utils/jsUtils.js';
 import { serializeToBuffer } from '../../utils/serialize.js';
 import { VK_TREE_HEIGHT } from '../constants.js';
 import { UInt32, UInt8Vector } from '../shared.js';
@@ -9,6 +8,7 @@ import { VerificationKey } from '../verification_key.js';
 import { KernelCircuitPublicInputs } from './public_inputs.js';
 import { makeEmptyProof } from './private_kernel.js';
 import { BufferReader } from '@aztec/foundation/serialize';
+import { assertLength } from '@aztec/foundation/utils';
 
 /**
  * Data of the previous kernel iteration in the chain of kernels.
