+++ yarn-project/circuits.js/src/structs/kernel/private_kernel.ts
@@ -1,5 +1,5 @@
 import { EthAddress } from '@aztec/foundation/eth-address';
-import { FieldsOf, assertLength } from '../../utils/jsUtils.js';
+import { FieldsOf, assertLength } from '@aztec/foundation/utils';
 import { serializeToBuffer } from '../../utils/serialize.js';
 import { PrivateCallStackItem } from '../call_stack_item.js';
 import { CONTRACT_TREE_HEIGHT, FUNCTION_TREE_HEIGHT, PRIVATE_CALL_STACK_LENGTH } from '../constants.js';
