+++ yarn-project/circuits.js/src/structs/rollup/base_rollup.ts
@@ -1,4 +1,4 @@
-import { assertLength, FieldsOf } from '../../utils/jsUtils.js';
+import { assertLength, FieldsOf } from '@aztec/foundation/utils';
 import { serializeToBuffer } from '../../utils/serialize.js';
 import {
   CONTRACT_TREE_HEIGHT,
