+++ yarn-project/foundation/package.json
@@ -27,6 +27,7 @@
     "./sleep": "./dest/sleep/index.js",
     "./timer": "./dest/timer/index.js",
     "./transport": "./dest/transport/index.js",
+    "./utils": "./dest/utils/index.js",
     "./wasm": "./dest/wasm/index.js",
     "./bigint-buffer": "./dest/bigint-buffer/index.js"
   },
